Changes based on pr feedback
