--- src/overlay-root/positioner.ts
@@ -1,66 +1,0 @@
-/*
-Copyright 2018 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-
-import calculatePosition, { PositionResult } from './calculate-position';
-
-const defaultOptions: Options = {
-    containerPadding: 10,
-    crossOffset: 0,
-    flip: true,
-    offset: 0,
-    placement: 'left',
-};
-
-interface Options {
-    containerPadding: number;
-    crossOffset: number;
-    flip: boolean;
-    offset: number;
-    placement: string;
-}
-
-export default class Positioner {
-    private target: HTMLElement;
-    private relativeElement: HTMLElement;
-    private container: HTMLElement;
-
-    public constructor(
-        target: HTMLElement,
-        relativeElement: HTMLElement,
-        container: HTMLElement
-    ) {
-        this.target = target;
-        this.relativeElement = relativeElement;
-        this.container = container;
-    }
-
-    public calculatePosition(options: Options): PositionResult {
-        const positionOptions = { ...defaultOptions, ...options };
-
-        return calculatePosition(
-            positionOptions.placement,
-            this.relativeElement,
-            this.target,
-            this.container,
-            positionOptions.containerPadding,
-            positionOptions.flip,
-            this.container,
-            positionOptions.offset,
-            positionOptions.crossOffset
-        );
-    }
-
-    public updatePosition(options: Options): PositionResult {
-        const position = this.calculatePosition(options);
-        return position;
-    }
-}
