+++ tsconfig.json
@@ -9,7 +9,7 @@
         "strict": true,
         "noUnusedLocals": true,
         "target": "es2017",
-        "types": ["mocha", "chai", "sinon", "karma-chai", "node"],
+        "types": ["mocha", "chai", "sinon", "karma-chai"],
         "typeRoots": ["./node_modules/@types", "./types"],
         "lib": ["es2017", "dom", "dom.iterable"],
         "outDir": "lib"
